refactor(ssh): extract marker protocol into its own module
The marker-based stdout filtering used to drive single-command SSH sessions lived inline in
the command handler, which made it impossible to cover with unit tests — every path went
through a real ssh subprocess and `process.stdout`/`process.stderr`. Lift the logic into
`src/models/ssh-marker-protocol.js` with the child process, marker, and output streams
injected, so the protocol (marker handshake, pre-marker buffering, failure-mode flush) can
be exercised in isolation.

No behavior change for the ssh command itself; the handler now just spawns ssh and hands
the subprocess to `runMarkerProtocol`.
